+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## UNRELEASED
+- ✨ Added `react` eslint config.
 
 ## v3.2.1 - 2024-10-04
 - 🐛 Added missing vue-tsc package.